pyproject.toml
[tool.poetry]
name = "reinforcement_learning_pipeline"
version = "0.1.0"
description = ""
authors = ["oren-neumann-motorai <oren.neumann@motor-ai.com>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
carla = "0.9.15"
numpy = "<2.3"
absl-py = "^2.3.1"
antlr4-python3-runtime = "^4.13.2"
astor = "^0.8.1"
cachetools = "^6.1.0"
certifi = "^2025.8.3"
charset-normalizer = "^3.4.3"
cloudpickle = "^3.1.1"
cycler = "^0.12.1"
dictor = "^0.1.12"
easydict = "^1.13"
ephem = "^4.2"
Farama-Notifications = "^0.0.4"
fonttools = "^4.59.1"
future = "^1.0.0"
gast = "^0.6.0"
google-pasta = "^0.2.0"
grpcio = "^1.74.0"
gym = "^0.26.2"
gymnasium = "^1.2.0"
gym-notices = "^0.1.0"
h5py = "^3.14.0"
idna = "^3.10"
importlib-metadata = "^8.7.0"
importlib-resources = "^6.5.2"
jax-jumpy = "^1.0.0"
joblib = "^1.5.1"
Keras-Applications = "^1.0.8"
Keras-Preprocessing = "^1.1.2"
kiwisolver = "^1.4.9"
laspy = "^2.6.1"
Markdown = "^3.8.2"
MarkupSafe = "^3.0.2"
matplotlib = "^3.10.5"
mpi4py = "^4.1.0"
nvidia-cublas-cu11 = "^11.11.3"
nvidia-cuda-nvrtc-cu11 = "^11.8.89"
nvidia-cuda-runtime-cu11 = "^11.8.89"
nvidia-cudnn-cu11 = "^9.10.2"
oauthlib = "^3.3.1"
packaging = "^25.0"
pandas = "^2.3.1"
pexpect = "^4.9.0"
pillow = "^11.3.0"
protobuf = "^6.32.0"
ptyprocess = "^0.7.0"
pyasn1 = "^0.6.1"
pyasn1-modules = "^0.4.2"
pydot = "^4.0.1"
pygame = "^2.6.1"
pyparsing = "^3.2.3"
python-dateutil = "^2.9.0"
py-trees = "^2.3.0"
pytz = "^2025.2"
PyYAML = "^6.0.2"
requests = "^2.32.5"
requests-oauthlib = "^2.0.0"
rsa = "^4.9.1"
shapely = "^2.1.1"
Shimmy = "^2.0.0"
simple-watchdog-timer = "^0.1.1"
six = "^1.17.0"
smmap = "^6.0.0"
stable-baselines3 = "^2.7.0"
tabulate = "^0.9.0"
tensorboard = "^2.20.0"
tensorboard-data-server = "^0.7.2"
tensorboard-plugin-wit = "^1.8.1"
termcolor = "^3.1.0"
transforms3d = "^0.4.2"
typing-extensions = "^4.14.1"
urllib3 = "^2.5.0"
Werkzeug = "^3.1.3"
wrapt = "^1.17.3"
zipp = "^3.23.0"
lanelet2 = "^1.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
